Split the .almd-to-Rust compile pipeline out of main.rs into compile_driver.rs
